requirements.txt
numpy
jax>=0.3.3
gym>=0.23.1
pygame>=2.1.2
open_spiel>=1.1.0
absl-py>=1.0.0
dm-tree>=0.1.6